tutorial-02/pyproject.toml
[project]
name = "tutorial-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.0",
    "strands-agents>=1.29.0",
    "strands-agents-tools>=0.2.22",
]
